api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
clerk-backend-api==1.0.0
